Combine openshift-tests and openshift-tests-upgrade suite
We differentiate tests on suite and name, because I imagined a scenario
where two entirely different test suites create a generically named test
like "TestClusterCreate" and we don't want to conflate them.  However,
openshift-tests and  openshift-tests-upgrade are definitely the same
tests, and it makes sense to combine them. We already get the "upgrade"
view from the variants.