Tag QE tests as LEVEL0
This collects all the variants a test is seen in, and then assigns any
job appearing in automated-release jobs as LEVEL0.

LEVEL0 tests are baseline QE tests. They run in many jobs, but
automated-release jobs *only* run LEVEL0 so it's a way we can identify
them. LEVEL0 and other metadata gets stripped from JUnits before
publishing, so it's the only way we can currently gather that data.